--- src/lib/components/types/ClassDetails.ts
@@ -1,20 +1,0 @@
-export type ClassDetails = {
-    classDay1: string
-    classTime1: string
-    classDay2: string
-    classTime2: string
-    meetingLink: string
-    gradeRecommendation: string
-    course: string
-    submitting: boolean
-    meetingTimes: Date[]
-    datesHeld: Date[]
-    feedbackCompleted: boolean[]
-    classStatuses: string[]
-    instructorFirstName: string
-    instructorLastName: string
-    instructorEmail: string
-    otherInstructorEmails: string
-    classCap: number
-    online: boolean
-  }
